requirements.txt
numpy
pandas
scikit-learn
torch
torchvision
torchaudio
opencv-python~=4.12.0.88
ffmpeg-python~=0.2.0
moviepy
librosa
pytube
streamlit~=1.50.0
plotly
matplotlib
tqdm
requests

yt-dlp~=2025.10.22